+++ src/xml/xml_native_writer.cc
@@ -142,6 +142,10 @@ void mjXWriter::OneFlex(XMLElement* elem, const mjCFlex* flex) {
   string  text;
   mjCFlex defflex;
 
+  if (flex->elastic3d) {
+    throw mjXError(0, "stable Neo-Hookean elasticity is mjSpec-only and cannot be written to MJCF");
+  }
+
   // common attributes
   WriteAttrTxt(elem, "name", flex->name);
   WriteAttrTable(elem,